POC: Use Axum web server for RESTful api
